expander done?

+++ libft/srcs/str/ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: tcensier <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/07 12:40:24 by tcensier      #+#    #+#                 */
-/*   Updated: 2022/10/07 16:04:04 by tcensier      ########   odam.nl         */
+/*   Updated: 2023/09/20 16:59:18 by rares         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
-#include <stddef.h>
 
-size_t	ft_strlen(const char *str)
+
+int	ft_strlen(const char *str)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (str[len])
